Formating transformer code
